reorganized files so that store is in "models" folder
